Changes in Finance tool
Former-commit-id: d9a7dca64e4710549636518883ed469be1c69ace

+++ src/com/vimukti/accounter/web/client/imports/ImportAction.java
@@ -16,11 +16,11 @@ public class ImportAction extends Action {
 	private int importType;
 	private List<ImportField> importerFields;
 	private String fileID;
-	private int noOfRows;
+	private double noOfRows;
 
 	public ImportAction(List<ImportField> importerFields,
 			Map<String, List<String>> data, int type, String fileID,
-			int noOfRows) {
+			double noOfRows) {
 		this.columnData = data;
 		this.importType = type;
 		this.importerFields = importerFields;
